rename module to comscore and comscoresample rename package to com.bitmovin.player.integration
