+++ comscore/build.gradle
@@ -2,7 +2,7 @@ apply plugin: 'com.android.library'
 apply plugin: 'com.jfrog.artifactory'
 apply plugin: 'maven-publish'
 
-def packageName = 'com.bitmovin.player.integrations'
+def packageName = 'com.bitmovin.player.integration'
 def libraryVersion = '0.9.0'
 
 android {
