+++ comscore/src/main/java/com/bitmovin/player/integration/comscore/ComScoreBitmovinAdapter.java
@@ -1,4 +1,4 @@
-package com.bitmovin.player.integrations.comscoreanalytics;
+package com.bitmovin.player.integration.comscore;
 
 import android.util.Log;
