+++ comscore/src/main/java/com/bitmovin/player/integration/comscore/ComScoreConfiguration.java
@@ -1,4 +1,4 @@
-package com.bitmovin.player.integrations.comscoreanalytics;
+package com.bitmovin.player.integration.comscore;
 
 public class ComScoreConfiguration {
     private String publisherId;
