+++ comscore/src/main/java/com/bitmovin/player/integration/comscore/ComScoreMediaType.java
@@ -1,4 +1,4 @@
-package com.bitmovin.player.integrations.comscoreanalytics;
+package com.bitmovin.player.integration.comscore;
 
 public enum ComScoreMediaType {
     LONG_FORM_ON_DEMAND,
