+++ comscoresample/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 27
     defaultConfig {
-        applicationId "com.bitmovin.player.integrations.bitmovincomscoreanalyticsexample"
+        applicationId "com.bitmovin.player.integration.bitmovincomscoreanalyticsexample"
         minSdkVersion 16
         targetSdkVersion 27
         versionCode 1
@@ -23,9 +23,9 @@ android {
 }
 
 dependencies {
-    implementation project(path: ':comscoreanalytics')
+    implementation project(path: ':comscore')
 
-//    implementation 'com.bitmovin.player.integrations:comscoreanalytics:0.9.0'
+//    implementation 'com.bitmovin.player.integration:comscore:0.9.0'
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.android.support:appcompat-v7:27.0.1'
