+++ comscoresample/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.bitmovin.player.integrations.bitmovincomscoreanalyticsexample">
+    package="com.bitmovin.player.integration.ComScoreExample">
     <uses-permission android:name="android.permission.INTERNET"/>
     <!--<uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>-->
     <!--<uses-permission android:name="android.permission.ACCESS_WIFI_STATE"/>-->
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.bitmovin.player.integration.ComScoreExample.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
