+++ comscoresample/src/main/res/mipmap-xhdpi/ic_launcher.png
