+++ comscoresample/src/main/res/values/colors.xml
