+++ comscoresample/src/main/res/values/strings.xml
