+++ comscoresample/src/main/res/values/styles.xml
